+++ HISTORY.md
@@ -3,6 +3,8 @@ v0.8 - 2014/07/18
 
 * [breaking change] channel.subscribe returns nil (not raise an error) when trying to subscribe to your own channel
 * Add all the status fields to Video (upload status, failure reason, rejection reason, scheduled time, license, embeddable, public stats viewable)
+* Add content_owner.claims to list the claims administered by a content owner.
+* Allow content_owner.claims to be chained with .where, such as in account.videos.where(q: 'query')
 
 v0.7 - 2014/06/18
 -----------------